Complete route for login and sign up
Use case 1 and 2 should be functional after this commit. Modified
model.py to support additional relationships between user and interests
as well as user and language plus skills. The About and T&C pages
should also be functional.

We'll need to add a helper function to force login on certain routes.